pyproject.toml
# Copyright (c) 2026 Guy's and St Thomas' NHS Foundation Trust & King's College London
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#     http://www.apache.org/licenses/LICENSE-2.0
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.
#

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flip-utils"
dynamic = ["version"]
description = "FLIP - Federated Learning and Interoperability Platform library built on NVIDIA FLARE"
readme = {file = "docs/overview.rst", content-type = "text/x-rst"}
requires-python = ">=3.12"
license = "Apache-2.0"
authors = [
    { name = "AI Centre for Value Based Healthcare" },
    { name = "Guy's and St Thomas' NHS Foundation Trust" },
    { name = "King's College London" },
    { name = "A. Triay Bagur", email = "alexandre.triay_bagur@kcl.ac.uk" },
    { name = "P. Wright", email = "p.wright@kcl.ac.uk" },
    { name = "R. Garcia-Dias", email = "r.gaciadias@gmail.com" },
    { name = "V. Fernandez", email = "virginia.fernandez@kcl.ac.uk" },
]
maintainers = [
    { name = "A. Triay Bagur", email = "alexandre.triay_bagur@kcl.ac.uk" },
    { name = "P. Wright", email = "p.wright@kcl.ac.uk" },
    { name = "R. Garcia-Dias", email = "r.gaciadias@gmail.com" },
    { name = "V. Fernandez", email = "virginia.fernandez@kcl.ac.uk" },
]
keywords = ["federated-learning", "medical-imaging", "nvflare", "flip"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
dependencies = [
    "boto3>=1.38.31",
    "nvflare==2.7.1",
    "pandas>=2.3.0",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/londonaicentre/flip-fl-base"
Repository = "https://github.com/londonaicentre/flip-fl-base"
Documentation = "https://flip-fl-base.readthedocs.io/en/stable"
Issues = "https://github.com/londonaicentre/flip-fl-base/issues"

[project.optional-dependencies]
# Full dependencies for running FL jobs (includes ML frameworks)
full = [
    "einops==0.8.1",
    "gdown==5.2.0",
    "lpips==0.1.4",
    "monai>=1.5.1",
    "nibabel>=5.3.2",
    "pydicom>=3.0.1",
    "pytorch-ignite>=0.4.11",
]

[tool.hatch.version]
path = "flip/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["flip"]

[tool.hatch.build.targets.sdist]
include = [
    "/flip",
    "/README.md",
    "/LICENSE.md",
    "/NOTICE.md",
    "/pyproject.toml",
]
exclude = [
    "*.pyc",
    "__pycache__",
    ".coverage",
    "*.coverage",
]

[dependency-groups]
dev = [
    "huggingface-hub>=1.6.0",
    "matplotlib>=3.10.7",
    "mypy>=1.18.2",
    "pyqt5>=5.15.11",
    "pytest>=8.4.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.2",
    "tensorboard>=2.20.0",
    "torchvision>=0.24.1",
    "types-requests>=2.32.4.20260107",
]

[tool.ruff]
line-length = 120

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PT']

[tool.pytest.ini_options]
testpaths = ["tests/unit"]
# Exclude subdirectories that have their own pyproject.toml and dependencies
norecursedirs = ["fl_services", "fl-base", ".venv", "workspace", ".test_*"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true

# Exclude patterns to avoid duplicate module issues
exclude = [
    "^tests/",
    "^fl_services/",
    "^workspace/",
    "^\\.test_",
    "^\\.venv/",
]

# Each app has its own 'custom' module, so we need to treat them separately
# by using explicit-package-bases with the namespace_packages option
explicit_package_bases = true
namespace_packages = true

# Ignore duplicate module errors for these specific patterns
[[tool.mypy.overrides]]
module = [
    "custom",
    "custom.*",
    "utils",
    "utils.*",
]
ignore_errors = true
